Enhance call stack traversing in getInfo
With the old implementation code like

Function testHelper_IGNORE()
	FAIL()
end

Function myTestCase()
	testHelper_IGNORE()
End

did not report the correct call site.
This was due to the assumption of a fixed call chain order.

Now we traverse the call stack from bottom up and report the first
function not in one of the unit testing procedure files.