Enhance error checking for override test hooks
Previously we searched for all correctly named functions with 1 parameter.
This created hard to find bugs if your hook function had a
different signature as it was silently ignored.

Now we don't limit the search according to the number of
parameters but check that each hook function has the correct number of
parameters and correct return type.