Add new assertions types for strings
Checking that a string is non null or non empty was not possible out of
the box. Add assertions *_NON_EMPTY_STR and *_NON_NULL_STR for testing the
assertions that a string is not null or not empty.

For convenience a null string is not considered non-empty but an empty string
non-null.

Signed-off-by: Thomas Braun <thomas.braun@byte-physics.de>