Added JUNIT support
Writes out the results of multiple TestSuites to a XML file in JUNIT
format.
added unit-testing-junit.ipf with the junit specific functions
The junit module holds Structures that are used in the main loop to
gather TestCase and TestSuite information. After each Case/Suite partial
XML output is buffered in a string.