GetBaseFilename: Avoid locale dependent timestamp
Since the dawn of this project we use parameter 1 for Secs2Time.

This uses according to the documentation a locale dependent way of
representing time. And on an english windows this creates names like
XXX_2016-9-10_11-39-19 AM.YYY for all three log file types.

Avoid that by using 3 which selects military time for Secs2Time.