JUNIT support: Make testcase classname unique for successive runs
Bamboo, which can be used as JUNIT consumer, does currently require that
the combination of classname + name is unique for all run tests.

If now a user runs the same tests multiple times, bamboo will one use
the last classname + name result.

To workaround that bug we introduce a run_count variable which tracks
how often the UTF was run in one Igor Pro session.

[1]: https://jira.atlassian.com/browse/BAM-14692