Catch aborts in test hook override functions
Until now code like

  #pragma TextEncoding = "UTF-8"
  #pragma rtGlobals=3
  #pragma ModuleName=WB_Testing

  static Function TEST_CASE_BEGIN_OVERRIDE(testCase)
      string testCase

      Abort
  End

results in the following output

Start of test "Unnamed"
Entering test suite "UTF_WAvebuilder.ipf"
Entering test case "WB_Testing#WB_RegressionTest"

which shows that the Abort in the test hook override is not caught and
thus the test run is not finished. Also outputs like JUNIT or TAP would
be incomplete.

Introduce ExecuteHooks which takes care of catching aborts. It also
takes care of calling the builtin as well which makes the call sites
much nicer. If an user hook abort is caught we also set the abort flag
and thus finish ASAP.

For user aborts we don't properly cleanup, i.e. call the end hooks, as
that would be way too involved.