error checking for getTestCasesMatch
this unites the use of code passages where Test Cases are searched in
Procedure files.

* introduce error checking to getTestCasesMatch.

* case insensitive match for regexp and non-regexp returning
TC_NOT_FOUND if no match was found for a given list of test cases.

* Error checking itself is not touched, meaning that the same errors are
produced as before.

Note:

The error is fatal once in the beginning of RunTest, where the list of
all TestCases is retrieved. Since the same function is called later on,
there will be no remaining fatal error in the main loop.