EQUAL_WAVE_WRAPPER: Make modes wave signed
The documentation of EqualWaves states that you can also pass in -1 as
mode and that checks all.

This does not work for EQUAL_WAVE_WRAPPER as we put the mode into an
unsigned wave and Igor's behaviour on signed -> unsigned conversion
currently clips that to zero.

Let's use a signed integer wave to support that as well.