Added support for test code running in background functions
The logic for this feature is as follows:

The UTF has now an own background task that is able to monitor the state of
user background tasks. The UTF user can register his test background tasks
to be monitored. The UTF will drop after the current test case function to the
Igor command line and resume test case execution when the user background task
stops.
To accomplish this RunTest has to save the complete current state of the UTF.
All relevant variables for RunTests local state have been encapsulated in a
structure. The function SaveState and RestoreState is used to save and to
restore UTFs local state.
For the autorun functionality the Igor Quit condition has changed. Previously
Igor quit if the run() function returned. This was incompatible with the
background task feature and is changed that Igor quits when RunTest finishes.
This also means that multiple calls to RunTest() in run() when the AUTORUN_FULL
condition is set is NOT supported any more.

Added example11-background.ipf to docu/examples to show how
test with background functions are setup.

Added example12-background-using-hooksipf to docu/examples to show how
test with background functions registered from a begin hook are setup.

Added example15-background_caught_abort_and_require.ipf to internal_dev.
It shows the behavior for the special case when a test case has an Abort and
when a non fulfilled REQUIRE is encountered, where execution of further tests
should stop.

Added IM_example14-background_reentry_module.ipf to internal_dev.
It shows how the procedure names should be referenced when using an IM and
regular Module in combination.