Reentry function supports now multi data input of same format as origin tc
moved the function signature check for multi data test cases to
GetFunctionSignatureTCMD

With the introduction of reentry functions of the same type as the
originating test case all Functions of type name_REENTRY([arg]) are
filtered out in getTestCaseList as well.

Added logic for multi data reentry function.

FuncRef is checked for multi data reentry function and abort on incompatible
data types is introduced. The abort logic uses the fact that FUNCREF that are
null call the prototype function. All proto type function have been extended to
throw a RTE, such that RunTest jumps to the catch branch and that single exit
path is used.

The local reentry is now reset at the end inside the while loop for the multi
data test cases.

Added example14-multi-test-data-with-background.ipf to internal_dev