Add Line Number Of Failing Test Case Line
All failing test cases were pointing to the line where the assertion
happened. This led to unreadable test case failures if the fail was
happening in sub-function.

With this changes, the caller stack is traversed from the last to the
first function that does not belong to the "unit-testing-framework"
namespace. Both line numbers are outputted to the cli. The failing line
is displayed as usual and the calling line is added in normal brackets.

Closes #75