Case of "no assertion" gets handled as error
Instead of a direct print to console in case of no assertion, ReportResults
is now called. The error counter is increased and the message is displayed
via PrintFailInfo.
To get the correct message GetInfo needed to get another return-case,
and the case "Assertion failed in unknown location." now only triggers if
the assertion counter is unequal 0.