New option debugMode for RunTest
The new option debugMode was inserted, which allows bit-input to select
a certain debugger mode. allowDebug works the same as before.
TestBegin and TestEnd now need information on both allowDebug and debugMode,
which is why both are connected as one variable and passed to both
functions.
Global variable igor_debug_assertion added to store the information on
debugModes' 4th bit.