GetFullFunctionName bugfix to append module only when needed
Now the return only contains the module if the function is declared static.

Before the module was always added. As the output is only used by FuncRef,
it did not matter if a public function had the module name appended or not.
Static functions were correctly addressed and priorized over public
functions with the same name.

Now the behavior is the same for static functions. For public functions
without module specification FuncRef still takes only the public
function.

Furthermore the code was cleaned up to apply better to our coding
conventions.