GetFunctionTagWave: Make the tag detection less buggy
Since the introduction of GetFunctionTagWave 935f7554 (Added utility functions to handle function tags centrally, 2020-05-26) cases like

 #pragma TextEncoding = "UTF-8"
 #pragma rtGlobals=3 // Use modern global access method and strict wave
 access.
 #pragma rtFunctionErrors=1
 #pragma ModuleName=TEST

 // UTF_TD_GENERATOR ABCD
 Function CheckIfAllControlsReferStateWvABCDEFGH([str])
   string str

   End

   Function/WAVE ABCD()

     make/t/free data

         return data
         End

resulted in

 Could not find data generator specification for multi data test case
 CheckIfAllControlsReferStateWvABCDEFGH. The tag UTF_TD_GENERATOR was not found.
   Error 24 in getTestCasesMatch: No test case found
     Error: A test case matching the pattern ".*" could not be found in test suite(s) "Proc0;".

The reason is that the search for the tags was buggy. The algorithm
tried to find the function keyword and stopped after that. But the
function keyword detection found also rtFunctionError.

The new algorithm calls ProcedureText twice, once with context and once
without and then removes the function body, so that only the comment
lines above are present. And it then traverses the comment lines back to
front. With passing -1 for the context in ProcedureText we get
everything above until the first blank line.