Fix for enforce of Datagenerator function to be static in Module.
In commit 98f58595e5dfd130d7b1d8e44f728724e70bca6e it was introduced that the
data generator function was enforced to be static when it was present in the
same procedure file.

However when the function is static then the procedure must have a Module
specification. Thus, the logic was changed that it gives an error only if
a static function was found without Module specification.
Also without Module specification (for non-static functions) only the function
name is returned.