At least one assertions in test case is enforced without assertions in hooks
Previously assertions that were called from hooks were counted as well,
that was not intended and led to cases where test cases with no assertion
did not error out when a related hook called an assertion.

Now the change of the assertion counter is monitored over a test case run,
thus assertions in hooks do not contribute to the test case assertion count.

Two new functions in the call chain logic were added:
BeforeTestCase and AfterTestCase
These functions are directly called before and after the test case not
including any user hook.