AreWavesEqual creates descriptive message in WAVE_DATA mode
If AreWaveEqual detects a wave difference through EqualWaves in mode WAVE_DATA
a descriptive message with up to ten locations in the waves is generated with
differences.
When the difference are found for the generation of the message the algorithm
follows EqualWaves behavior.
In addition to the wave names a table output is created with dimension, labels
and values.

Also added an internal_dev/EqualWaves_Examples.ipf with failing tests and
example output from various wave differences.

Limitations:
- Values from complex INT64/UINT64 waves are printed incorrect.
- In IP9 pointer values are only printed if there is no pending RTE present
  if there is a pending RTE then "_free_" is printed.
- The extended string value output checks only against
  C0, C1, Bi-Directional Control characters and 0x007F (DEL).
  It also does not take into account if the selected font has actual support
  for displaying specific characters. Unsupported printable characters will be
  shown as e.g. empty boxes.