Introduced typed waves for multi data test cases
In addition to the generic WAVE type multi data test cases also accept now
WAVE/T, WAVE/WAVE and WAVE/DF specific types.
The associated data generators must return waveRef waves where each element
must have the fitting type.

Added examples for all three new typed waves to DataGenSpecificWaveTypes.ipf
in internal_dev.

The MD test case data type setup is not internally buffered for each test case
found. This means that for the actual call to the test case the function
template is determined from the wave type of the element of the data generator
wave.
So for a data generator that returns a wave reference wave with text waves as
elements a test case with WAVE and with WAVE/T is valid. This applies to
reentry functions as well. See also the last example in
DataGenSpecificWaveTypes.ipf.
There is no priority applied as with a test case name only one declaration is
connected.