Fix for RTE cleanup in AreWavesEqual blocking TC RTE evaluation
A fix for invalid parameters for a EqualWaves call in the IUTF was introduced
in 28f7fc37221c117ae3d4a6e5baddd4dd549cc4bb

This created the problem that an RTE that happens in a test case is cleared
at this point in the IUTF and then the RTE gets not visible in the RTE
evaluation. Thus it gets ignored and does not raise an error.

Also specific to IP6 EqualWaves does not raise an RTE but returns NaN instead.

The fix:
- Removed the EqualWaves monitoring with AbortOnRTE.
- Added manual parameter check for mode and tol in EQUAL_WAVE_WRAPPER,
  that prevents an RTE to be raised by EqualWaves.

Mode is checked in EQUAL_WAVE_WRAPPER for validity to abort early on
invalid modes. mode must have at least one valid bit set.
Tolerance tol must be 0 or positive and not NaN.

Added RTE tests for EQUAL_WAVE_WRAPPER to VTTE with invalid mode, tol