With JUnit enabled a <failure> tag for each failed assertion is writen
Previously only the last failure message appeared in the JUnit output.
Now for each failed assertion a <failure> or <error> tag is written,
depending if the test case aborted or not.

Also <system-err> output was extended accordingly to include all
fail messages.

The xml file with multiple failure tags was checked against the
reference xsd schema file and it validates (using xmllint)