fix for too long status message in IP6
In IP6 the string length for sprintf, printf is limited to 999 escaped
characters.
This generated an RTE in the functions
DebugOutput
PrintFailInfo
UTF_PrintStatusMessage

Generally we do not want to have the status message which is the main output
message to history to be cut short.

As in the first two functions the string was built from strings the sprintf
was replaced by string addition that does not suffer from the limit.

In UTF_PrintStatusMessage the output to the history was replaced with
print instead of printf for IP before 9. print does not suffer from the printf
limitation. The maximum line length before the first automatic line break was
set to the maximum the earlier IP versions offer.

For the fprintf the length was limited for IP8. For that
UTF_Utils#PrepareStringForOut got a new optional parameter to override the
maximum length returned. It is set to the maximum length IP8 offers.