Add GetWavePointer
This is a generic function to return the memory address of a wave which
works on all Igor Pro versions. Unfortunately the version checking is
still necessary as IP6 does not support Multithread on text waves. But
also IP6 does not create a RTE if we get the wave pointer via our wave
reference wave.