Code Cleanup for handling of allowDebug parameter
allowDebug is now handled through debugMode internally.

debugMode sets the state of the debugger while a testcase runs,
if debugMode is set.

Thus, allowDebug just means that the debugger state while the IUTF is running
is the same as outside the IUTF. This is set now accordingly.

Also added Tests to track the Debugger Settings through a test run.