bugfix: lines starting with elseif got instrumented as lines with else
Since commit de9e739cd0492ddd2432c01a53a658767494197e

Lines starting with elseif were instrumented the same as lines with else,
where a Z function was added in the following line. The condition of the elseif
line was not replaced.
The cause was that the mentioned commit changed the handling of lines starting
with else to the described behavior. However, lines with keys, such as else,
were checked before conditions like elseif/if. Since, "elseif" also starts with
"else" it was misidentified as "else".

The solution is to change the priority order for adding the Z_ function.
Now conditions from lineStartZReplaceKeys are checked first.