Bugfix: Various.pxp was not executing Tests in Various.ipf in IP9
In IP9 the tests in Various.ipf were called with instrumentation.
In the same function RunTest was called again for the Debugger tests. Since
the function did not drop to the commandline no recompile/reentry could
happen and the actual tests from Various.ipf did not execute.

Moved the instrumentated test now to the end to be executed as last test.