Fix for stdout output of status messages length for IP9
Due to a wrongly ordered ifdef check for the running IP version the console
output of status messages was for IP9 cut short to the maximum allowed length
of IP8. This was no critical issue but unecessary as IP9 supports full length
strings through fprintf.
The order was fixed and now with IP9 full length status messages are outout to
stdout.