Added preprocessor endif and else to be counted after the line of appearance
The preprocessor endif and else were handled like regular code lines where
counting is done after the line executed. To be in line with the normal
if else endif behavior it was changed.
This applies only to conditionals inside functions.
No conditional replacement of the preprocessor if is possible as calling Z_
requires compiled state.