Add support for instrumentation of macros, procs and window macros
These changes suffer from limitations Igor Pro has for getting information
about its own procedure code. With functions it is possible to retrieve the
line in the procedure with FunctionInfo, which does not exist for Macros.
Thus, the tracing setup scans in a rather simple way the procedure text to find
the macro start.

Limitations:
Multipe macros with the same name in the same procedure file that are selected
with ifdef conditions on compile time are not supported as the tracing setup
can not determine which macro is actually executed.

Every procedure file that contains only macros must have a dummy function added
that allows the tracing setup to determine the procedure files file system
location though FunctionPath. Igor Pro offers no equivalent function for
macros.