Compare zero sized waves properly with set mode in EQUAL_WAVES_WRAPPER
The Igor Pro function "equalwaves" does return a RTE when the wave data should
be compared and one wave is of zero size.
For that we caught that condition beforehand and did a simple wave type check
independently of the mode set.

This was crude and let to confusion when actually using mode settings.

This commit fixes that behavior with two changes:
- The check for compatible wave types for Igor Pros equalwaves was adjusted
  to be applied only if at least one wave has data points.
  Igor Pros EqualWaves throws an RTE when Textual and Numerical waves should be
  compared.
- Otherwise UTF_Checks#AreWavesEqual is called with each single mode setting
  separately also for zero sized waves. In that function we handle the case
  mode == WAVE_DATA manually and do no call Igor Pros EqualWaves.