Require optional parameter for test methods
In Igor 6 there is no distinction between required and optional
parameter when matching functions. Therefore some hooks are used as
test cases and the UTF will later throw an exception because no data
generator was attached.

With this change every function for a test case will be checked to have
only optional parameter.

Fix #266.