Documentation: Fix the number of tags
There were a few places in the documentation where it said that a
maximum of 4 tags per function are recognised.

This is deprecated since the introduction of MMD test cases. Now all
lines up to the previous function can have tags.

This change reflect this new behavior in the documentation.

Close #269.