Enhance String Diff representation in wave diffs
Before the diff window for text waves showed the full strings in
hexadecimal format if the entry is different in both waves.

With this change the window will only show the start of the difference
and up to a maximum of 10 characters left and right. All displayable
characters in ASCII are shown as normal and the other letters are shown
in readable notation (the name of the character or hexadecimal form).

The window will also show which line, column and position in the full
string where the first difference happen. There is no special handling
of strings that are only one line.

This commit also includes an example experiment file in
"internal_dev/TestTextWaveDiff".

Close #274.