Enable size only checks for different typed waves
Before checking any wave property the unit testing framework will always
check if the types are compatible. This can yield to unintended behavior
when for example only the sizes should be checked and the waves are
intended to be different typed.

With this change the type check is restructured to enforce same typed
waves only for WAVE_DATA checks on non empty waves.

Close #199.