Allow unsaved experiment files.
Before any experiment had to be saved on the disk (it is allowed to have
unsaved changes). This was a requirement because the IUTF needs a set
home path to save data for JUnit, TAP and tracing

With this change it will always check if the current test run has any
feature activated that requires a local directory on the disk. If so it
will perform the usual check. If not the user can continue their work
with a completely unsaved experiment. This is useful when doing small
tests in Igor Pro and UTF.