Enhance error reporting in CHECK_WAVE
Makes the output of the wave type more user understandable. This shows
the specific type as a string which contains the same names as the flags
that are used to specify the type in CHECK_WAVE.

This commit also introduced NON_NUMERIC_WAVE as a new wave minor type
constant for all non numeric waves.

Close #220.