EQUAL_STR_WRAPPER: Make the default case sensitive
Although Igor Pro is in 99% of the cases case insensitive there are
exceptions. One of that is EqualWaves. As this does a binary comparison
for text waves it is case sensitive by default.

And that means that test cases like

static Function TestCaseSensitivityString()
	string outputOfFunc = "AbC"
	string expect = "abc"

	CHECK_EQUAL_STR(expect, outputOfFunc)
End

static Function TestCaseSensitivityWave()
	Make/FREE=1/T outputOfFunc = { "AbC" }
	Make/FREE=1/T expect = { "abc" }

	CHECK_EQUAL_WAVES(expect, outputOfFunc)
End

give

  Entering test case "mymodule#TestCaseSensitivityString"
  Leaving test case "mymodule#TestCaseSensitivityString"
  Entering test case "mymodule#TestCaseSensitivityWave"
  Assuming equality using mode WAVE_DATA for waves expect (0x085cfe4d60) and outputOfFunc (0x085cfe4d40); detailed: Text waves difference:
Wave1: expect (0x085cfe4d60)
Wave2: outputOfFunc (0x085cfe4d40)
Dimensions|Labels|Text        |
----------|------|------------|
[0]       |      |0:0:0> a b c|
          |      |0:0:0> A b C|
: is false. Assertion "CHECK_EQUAL_WAVES(expect, outputOfFunc)" failed in line 22, procedure "Proc0"
  Leaving test case "mymodule#TestCaseSensitivityWave"

which is puzzling as one would expect the same behaviour for both test
assertions and thus test cases.

We can't flip the case sensitivity switch for EqualWaves as there is none.
And a testing framework should also be as strict as possible, so we flip
the default for EQUAL_STR_WRAPPER to case sensitive.

Reported-by: @Garados007
Test-code-by: @Garados007