Abort test run if RTE ocurred in UTF
Abort the test run if some unhandled RTE still exists before
CallTestCase(). These errors are most likely a bug in UTF and leave the
UTF in an inconsistent state. It is more safe to stop the run at that
moment instead of continuing it.

Before this change any RTE that could happen before the run of the test
case are just ignored and thrown away.

Close #264.