AreVariablesClose: Prevent singularity
If AreVariablesClose is called with one of both arguments zero it will
trigger a singularity when executing "diff / abs(var)". This is an issue
in the used algorithm. The boost framework provided a solution for this
case as such as catching this special case and using a different
algorithm. [1]

With this fix AreVariablesClose can be called with any real values.

[1]: https://www.boost.org/doc/libs/1_77_0/libs/test/doc/html/boost_test/testing_tools/extended_comparison/floating_point/floating_points_comparison_theory.html

Close #216.