Feature: Check for free wave leaks in test cases
It is possible to write igor code that won't release free waves after
they are no longer used. This can conclude to big wave leaks and
therefore useless memory consumption.

With this commit the ability to check for these free wave leaks is
introduced. To enable this it is only required to set waveTrackingMode
in RunTest to UTF_WAVE_TRACKING_FREE, UTF_WAVE_TRACKING_LOCAL or
UTF_WAVE_TRACKING_ALL. This will track the number of created and
released waves during a test case run. If there is a difference this
will produce a CHECK error.

If the user wants to opt-out a single test case from this wave tracking
the test case as to be tagged with UTF_NO_WAVE_TRACKING.

The check for leaked waves is done after the test case end user hook to
enable the user to do some cleanup if necessary.

Close #200.