EnsureLargeEnoughWaveSimple: Use exponential growth
EnsureLargeEnoughWaveSimple checks and optionally resize a wave to
ensure that a given index is always inside the bounds. This function is
often used as a list of entries with an unknown size at start.

The previous approach to increase the size of the wave was to add a
chunk of fixed length. Which is good for small lists but is expected to
have a linear number of resizes when the number of entries becomes very
large.

Resizing a list is very costly because it will copy its content to the
larger list. Thus it is intentional to make the number of resizes as
small as possible. For this the new approach grows the waves exponential
with a logarithmic number of resizes. For small instances the number
of resizes are pretty similar to the linear approach but will become
very few if the number of entries become very large.

EnsureLargeEnoughWaveSimple will also check if all provided waves are at
least WAVECHUNK_SIZE long. This ensures that all given waves have at
least a sane capacity.

See: PR #289.