Output summary at the end of test run
Collect information when and where an assert failed and output the list
at the test run in the history area.

The output format where an assert is located is changed to be more
readable by the user. It also contains the name of the testcase and the
function that contains the assertion now.

Close #193.