FIX: error counter for test assertions
During some refactoring (813a2a3 (Combine PrintFailInfo and
TestCaseFail, 2022-11-14)) of the handling of internal error messages
the error counter got broken. It will increase the error always by one
for each error printing even it was an expected error.

This was due the property `incrErrorCounter` wasn't set to zero in
`ReportError` anymore. This is now fixed as such as an option for
`TestCaseFail` was introduced to enable setting `incrErrorCounter` to
zero again which is required for UTF assertions.

Close: #311.