Refactor TAP to use test result waves
The whole output and data collection of TAP is changed in that way that
is uses the data in the test results wave only. At the same time all
TAP hooks and global variables are removed.

TAP is no longer created during the test run and wil be created after
the test run finished.