Refactor reporting functions to be cleaner
The introduction of the test result waves changed the way how the result
of a test is stored or output to the history. For this have many output
functions been changed to accommodate these changes. The output
functions have been revised in this commit and refactored to be cleaner,
have less code and to be easier understandable.

1. The increment of the assertion error counter for the current test
    case have been moved to a single location in AddError().
2. The API of TestCaseFail() has been simplified and is easier to use
    for different error sources.
3. getInfo() has been moved from UTF_Basics to UTF_Reporting as its only
    user is in UTF_Reporting. The API is also simplified to return the
    partial stacktrace instead writing in to the result wave directly.
    This makes the usage easier.
4. PrintFailInfo() has been simplified.
5. ReportResults() has been simplified.
6. The API of EvaluateRTE() has been modified to make it easier to know
    when an error happened in context of a test case.