Refactor global AssertionInfo string to wave
This enables to store each single information and make it easier to
store them in the test result waves.

During refactoring the name is changed from AssertionInfo to InfoMsg to
prevent name clashing with existing igor experiments and to make it less
confusing for the developer as there are two test result waves with
similar name (TestAssertionResult and TestInfoResult).

The location of the getter and clear method has been moved to
utf-reporting as this file seems more fitting for this purpose.