GetWavePointer_Impl: Fix return memory address
We don't want to prefix it with 08 but instead we want to have 08 as
precision which must stand *after* the percent sign.

While at it we also leverage the power of # and capital X to get a capital
0X prefix.

Bug introduced in b7bb12a4 (AreWavesEqual creates descriptive message in
WAVE_DATA mode, 2021-03-04).