Fix: Show INFO() for CHECK_EQUAL_WAVE()
The function EQUAL_WAVE_WRAPPER() contains multiple calls to
EvaluateResults() which clear up the AssertionInfo state after its call.
This yield to this unexpected behavior if the first call cleared the
information and the second call need it for output.

To prevent this there are two options:

1. Call "return NaN" after the call to EvaluateResults()
2. Call EvaluateResults() with "cleanupInfo = 0" and call
   UTF_Basics#CleanupInfoMsg() at the end.

The feature INFO() was introduced in 73dc1b8 (Add INFO() UTF function
for assertion context, 2022-11-23) but didn't catch this special case
in CHECK_EQUAL_WAVE().

Close #368