Function Tags: Fix long function name bug
Function tags are stored for each function in a wave reference wave.
This is done by assigning the full function name as a dimension label to
easier access the function tags.

The problem here is that dimension labels are limited to 31 characters
in Igor 6 and 7 and 255 characters since Igor 8. The full function name
can be way longer than this limit and this error is silently ignored.
This can also yield to the bug that the same function tags can be
applied for multiple functions.

This problem is solved as all full function names are stored in a text
wave. At the same position in the wave reference wave are their function
tags stored.

Close #378