Function Tags: Cleanup function tags
Igor Pro 9 has limited the allowed character set for dimension labels
which yield to an error if UTF_FTAG_TAP_DIRECTIVE or
UTF_FTAG_TAP_DESCRIPTION tags are used.

This has been fixed as all unsupported characters from both tags has
been removed and are now stored correctly.

Also a single colon between the function tag and its value can be
optionally as a separator added. It is not required but will make the
code more cleaner and understandable.

Close #328