Refactor: Move md test related setup to own files
UTF_Basics contain a lot of functions that is related for the setup of
multi data and multi-multi data test cases, the data generators and the
function prototypes. These have been moved into their own files and
UTF_Basics and the new procedure files are now more slim and easier to
manage.

This is done in preparation of #353.