Remove unit-testing-tracing-auto.ipf in favor of TUFXOP
unit-testing-tracing-auto.ipf stored the list of instrumented procedure
files as a readonly threadsafe free wave. This file is autogenerated and
stored alongside UTF itself. This has some issues as it is difficult to
manage inside a versioning system and it provokes problems when
different experiments are tested at the same time.

The data is stored now in a global wave inside memory using TUFXOP and
accessible whenever it's needed. This data is no longer written to disk.

Close #314