Refactor: Debug helper into own file
UTF_Basics contained some helper function that are used for setting up
the Igor Debugger. These are moved to UTF_Debug. Also the debugger
related code in UTF_Hooks for TestBegin and TestEnd has been moved to
UTF_Debug.

The static debug methods have been additionally refactored in the
following way:

- the code for EnableIgorDebugger has been moved to SetDebugger
- the code for DisableIgorDebugger has been moved to SetDebugger
- the code for RestoreIgorDebugger has been moved to RestoreDebugger

This is done as the only consumer are located in UTF_Debug and this
reduces code and make it more understandable.