Use cached home path for accessing files
At the start of RunTest the current home directory of the experiment
file is fetched and cached in a global string. If for some reasons a
test case changed to current home directory (for example storing the
experiment file elsewhere) all UTF functions will still use the old
home directory.

Also the access to the current home path has been refactored as there is
now a convenient function in UTF_Utils_Paths. The function FileNotExists
and getUnusedFileName has been moved from UTF_Basics to UTF_Utils_Paths.
getUnusedFileName is now static as it was marked as undocumented and
only used locally.

Close #287