Fix: Long wave labels in DataGeneratorWaves
DataGeneratorWaves used the full function names as wave labels to find
its entries. This can lead to bugs when the full function name becomes
larger than the supported wave label size. This is fixed using a second
text wave to store the function names at the same index as the entries
in DataGeneratorWaves.

This is the same fix as in 0309a67 (Function Tags: Fix long function
name bug, 2023-01-18).

See: #390