Fix: Long wave labels in MMDataState
MMDataState used the full function names as wave label to find its
entries. THis can lead to bugs when the full function name becomes
larger than the supported wave label size. This is fixed using a second
text wave to store the function names at the same index as the entries
in MMDataState.

This is the same fix as in 0309a67 (Function Tags: Fix long function
name bug, 2023-01-18) and 92e2ec4 (Fix: Long wave labels in
DataGeneratorWaves, 2023-02-21).

Close #390