*_EQUAL_STR and *_NEQ_STR are now pass-by-value
The string paramater for WARN_EQUAL_STR, CHECK_EQUAL_STR,
REQUIRE_EQUAL_STR, WARN_NEQ_STR, CHECK_NEQ_STR and REQUIRE_NEQ_STR are
now pass-by-value. This allows simpler usage and unify the api for the
*_EQUAL_* and *_NEQ_* variants.

These assertions were originally designed to be pass-by-reference to
catch the edge-case when the user provide a null string. There is no
longer a need take this in account as there are *_NULL_STR and
*_PROPER_STR assertions that are designed to check for these edge cases.

Close #363