Fix: Order of test case execution
The order of test case execution is now fixed and IUTF now always
executes the test cases from the same test suite together. The execution
of the test suites are now in that order they are listed in the RunTest
call (or alphabetically if regex used).

The code of CreateTestRunSetup is now much simpler and contains less
nested loops.

Close #416