Fix: Rename igortest in tests/Tracing.pxp
e46d6a4 (Rename: Update experiment files, 2023-02-14) has updated the
project name in all experiment files, which doesn't include Tracing.pxp.
Instead is a copy added to the repository which was created as a result
of a test case run. To prevent this in the future the test file is added
to the local gitignore file.