Tracing HTML: Use marker wave
The tracing HTML output used to mark all lines of a function as
instrumented. This has been changed as it will now use the marker wave
which contains the correct list of instrumented lines.

With this change the output of the iutf_instrumentation_data.txt file is
now obsolete and therefore removed.

See #414