Add IM Unit tests experiment
This experiment allows to run test cases inside an independent module
with tracing enabled for all IUTF files. This experiment creates a JUnit
and Cobertura report at the end of the test run.

The IMUnitTests experiment replaces IMTest as it is more feature rich
and includes all tests of IMTest.

See #425