Fix function signature in cobertura output
The ReportGenerator concats the function name and the full function
name of all methods of the cobertura output. This yields to unreadable
output like "UserPrintFIUTF_Utils_Strings#UserPrintF".

It also seems that the signature part is meant for the function
arguments. The output has been changed so that the method name is now
the full function name and the signature is the constant string "()".