Fix: UnitTests main.ipf
Added auto cleanup after test run. This simplifies the development
environment as there is only a need to do a single call. This change has
no influence to the CI as the changed igor procedure files are ignored.
If keeping of instrumented procedure files are intended for debug
purposes the developer can comment the test run end hook.

debugMode is also changed to allowDebug to simplify the development
environment.