Fix assertion source in independent modules
When an assertion triggers it will output the assertion call including
its arguments to the history. This works by grabing the source code of
the test and extracting the line that call the assertion. The previous
implementation could not determine the source code if the called
procedure file was inside an independent module.