CI: fail execution if no JUnit file found
The new statement checks explicitly if JUnit files are generated at all
and fail if it doesn't find any. In the previous setup the test will
succeed even if no JUnit files are found. The culprit for this is the
last line with the exit statement.

Let's say there is no JUnit file written by the test:

1. Bash will see `tests/**/JU_*.xml` as a file pattern but doesn't find
   any files for it. Bash won't replace this pattern and keep it as it
   is.
2. The first grep fails read `tests/**/JU_*.xml` and create an empty
   output.
3. The second grep happily accepts the empty output. The number of lines
   that doesn't contain the zero error string are indeed zero and
   prints a zero to the output.
4. exit receives the zero as an argument and quits with a successfull
   state.

It is possible to set `set -o pipefail` globally to fail a pipe if any
step of the pipe fails but this is known to have unexpected behavior in
the Gitlab CI.

See #432